requirements.txt
aiobotocore==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.12.0
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
boto3==1.40.66
botocore==1.40.66
cffi==2.0.0
cfgrib==0.9.15.1
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
eccodes==2.42.0
findlibs==0.1.2
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
s3fs==2025.10.0
s3transfer==0.14.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
xarray==2025.10.1
yarl==1.22.0
